requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.6
    # via kombu
argon2-cffi==20.1.0
    # via -r requirements.in
asgiref==3.4.1
    # via django
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
celery==5.1.2
    # via -r requirements.in
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via argon2-cffi
charset-normalizer==2.0.4
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
decorator==5.0.9
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.6
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-ltree
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-editorjs-fields==0.2.2
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.4.0
    # via
    #   -r requirements.in
    #   djangorestframework-filters
django-import-export==2.5.0
    # via -r requirements.in
django-ltree==0.5.3
    # via -r requirements.in
django-ordered-model==3.4.3
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-filters
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-filters==1.0.0.dev2
    # via -r requirements.in
djangorestframework-simplejwt==4.7.2
    # via -r requirements.in
docutils==0.17.1
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
emoji==1.4.2
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
idna==3.2
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-yasg
ipython==7.26.0
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
itypes==1.2.0
    # via coreapi
jedi==0.18.0
    # via ipython
jinja2==3.0.1
    # via coreschema
kombu==5.1.0
    # via celery
markuppy==1.14
    # via tablib
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.2
    # via ipython
multidict==5.1.0
    # via yarl
odfpy==1.4.1
    # via tablib
openpyxl==3.0.7
    # via tablib
packaging==21.0
    # via drf-yasg
parso==0.8.2
    # via jedi
pep517==0.11.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.1
    # via -r requirements.in
pip-tools==6.2.0
    # via -r requirements.in
prompt-toolkit==3.0.19
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.1
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pycparser==2.20
    # via cffi
pydash==5.0.2
    # via -r requirements.in
pygments==2.10.0
    # via ipython
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
python-json-logger==2.0.2
    # via -r requirements.in
pytz==2021.1
    # via
    #   -r requirements.in
    #   celery
    #   django
pyyaml==5.4.1
    # via tablib
redis==3.5.3
    # via -r requirements.in
requests==2.26.0
    # via
    #   -r requirements.in
    #   coreapi
ruamel.yaml==0.17.10
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
six==1.16.0
    # via
    #   argon2-cffi
    #   click-repl
    #   django-ltree
sqlparse==0.4.1
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.0.0
    # via django-import-export
tomli==1.2.1
    # via pep517
trafaret==2.1.0
    # via -r requirements.in
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.6
    # via requests
uwsgi==2.0.19.1
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
yarl==1.6.3
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
